+++ src/component/FolderInfo/FolderInfo.js
