+++ src/containers/Nav/Nav.js
@@ -1,6 +1,6 @@
 import logoImg from "../../assets/img/logo.svg";
 import style from "./Nav.module.css";
-import NavProfile from "../../component/NavProfile";
+import NavProfile from "../../component/NavProfile/NavProfile";
 import { useLocation } from "react-router-dom";
 import clsx from "clsx";
